+++ Node.h
@@ -16,7 +16,7 @@ private:
     Player human;
     Player ai;
     int last_row, last_col, last_val;
-    int score;
+    float score;
     std::vector<Node> children;
 
 public:
@@ -30,9 +30,9 @@ public:
     void update_last_move(int row, int col, int val);
     void generate_children(bool is_user_turn);
 
-    int search_ab(int curr_depth, int max_depth, int is_user_turn, int alpha, int beta);
+    float search_ab(int curr_depth, int max_depth, int is_user_turn, float alpha, float beta);
     void get_best_move(int &row, int &col, int &val);
-    int eval();
+    float eval();
 };
