+++ global.h
@@ -12,6 +12,6 @@
 const int EMPTY_CELL = 0;
 const int HUMAN_CELL = 1;
 const int AI_CELL = 2;
-const int MAX = 1000000000;
-const int MIN = -1000000000;
+const float MAX = 1000000000.0;
+const float MIN = -1000000000.0;
 const int MAX_NODE = 100000000;
